requirements.txt
faster-whisper
pyannote.audio
torch
torchaudio
onnxruntime-gpu
imageio-ffmpeg
python-dotenv
transformers
accelerate
bitsandbytes
